The module "quantum_mod.py" was added. This module allows to perform calculations which can be very close to the QC. 1) statistical error is introduced 2) simulation of the errors taken from the real computer can be introduced (set "noise_off = False" in the input file). 3) The code can be launched on a REAL QC
The code is now launched in a different way. On input, as an argument,
it expects the input file with the parameters (only nontrivial keys are
explained):
Q_on = True
If the value is "True" than the calculations are performed in QC-like
manner with statistics. If, in addition, "noise_off = False" then the
simulation of noise is added.

WARNING! Launch of the code with 8 qubits and noise can take a lot
of time! Prelimilarly calculations without errors are recommended.

The key "Nrep = 10" sets the number of runs on a simulator or real QC.
Nruns = Nrep * 8192. The QISKIT limits the possible number of runs by
1`000`000.